solana-agents/code-generation/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
openai>=1.6.1
esprima>=4.0.1
requests>=2.31.0
